added calcutation of closest city. added string files.

+++ data-ipma/src/main/java/com/dfl/dataipma/mapper/GlobalIdsDtoToCityListMapper.kt
@@ -11,6 +11,11 @@ class GlobalIdsDtoToCityListMapper {
     }
 
     private fun globalIdToCity(globalIdDto: GlobalIdDto): City {
-        return City(globalIdDto.globalIdLocal, globalIdDto.local, globalIdDto.latitude, globalIdDto.longitude)
+        return City(
+            globalIdDto.globalIdLocal,
+            globalIdDto.local,
+            globalIdDto.latitude.toDouble(),
+            globalIdDto.longitude.toDouble()
+        )
     }
 }
